[refactor] Update import paths for ReactQueryProvider and useAppear; rename Movie App to Movie; remove unused enabled flags in hooks
