+++ src/main.tsx
@@ -2,7 +2,7 @@ import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
 import '@/index.css';
 import App from '@/App';
-import ReactQueryProvider from '@/shared/context/ReactQueryProvider';
+import ReactQueryProvider from '@/components/common/ReactQueryProvider';
 
 createRoot(document.getElementById('root')!).render(
   <StrictMode>
